need to fix file streaming

+++ src/routers/fileUpload.ts
@@ -8,9 +8,10 @@ import { Module } from '../models/Module';
 import { File } from '../models/File';
 
 import { auth } from '../middlewares/auth';
-import { mustBeClassOwner } from '../middlewares/userLevels';
+import { mustBeClassOwner, mustBeStudentOrOwner } from '../middlewares/userLevels';
 
 import { SendOnError } from '../utils/functions';
+import { modulePath } from '../utils/paths';
 
 const router = express.Router();
 
@@ -34,19 +35,24 @@ router.post('/:classId/:moduleId', auth, mustBeClassOwner, async (req, res) => {
       },
     });
 
+    let errored = false;
+
     const data: {[fieldname: string]: string} = {};
 
     busboy.on('file', (_fieldname, file, filename) => {
-      const filenameToSave = `${nanoid()}${path.extname(filename)}`;
-      const saveTo = path.join(__dirname, `../../../media/class/modules/${filenameToSave}`);
-      const stream = fs.createWriteStream(saveTo);
-      file.pipe(stream);
+      if (filename.match(/\.(png|jpeg|jpg|gif|pdf|doc|docx|xls|xlsx|ppt|pptx|txt|mp4|mkv|mov|wmv)$/i)) {
+        const filenameToSave = `${nanoid()}${path.extname(filename)}`;
+        const saveTo = path.join(__dirname, `../../../media/class/modules/${filenameToSave}`);
+        const stream = fs.createWriteStream(saveTo);
+        file.pipe(stream);
 
-      data.filename = filenameToSave;
-
-      file.on('error', () => {
-        fs.unlink(saveTo, () => null);
-      });
+        data.filename = filenameToSave;
+        file.on('error', () => {
+          fs.unlink(saveTo, () => null);
+        });
+      } else {
+        errored = true;
+      }
     });
 
     busboy.on('field', (fieldname, value) => {
@@ -54,6 +60,9 @@ router.post('/:classId/:moduleId', auth, mustBeClassOwner, async (req, res) => {
     });
 
     busboy.on('finish', async () => {
+      if (errored) {
+        res.status(400).send({ error: 'Check your file type or try again later' });
+      }
       data.moduleId = req.params.moduleId;
       const createdFile = await File.create(data);
       res.setHeader('Connection', 'close');
@@ -63,6 +72,14 @@ router.post('/:classId/:moduleId', auth, mustBeClassOwner, async (req, res) => {
     return req.pipe(busboy);
   } catch (e) {
     return SendOnError(e, res);
+  }
+});
+
+router.get('/:classId/:moduleId/:filename', auth, mustBeStudentOrOwner, async (req, res) => {
+  try {
+    res.sendFile(`${modulePath}/${req.params.filename}`);
+  } catch (e) {
+    SendOnError(e, res);
   }
 });
